requirements.txt
flask==2.0.1
pytesseract==0.3.8
Pillow==8.3.1
pandas==1.3.0
opencv-python-headless==4.5.2.54
numpy==1.21.0
openpyxl==3.0.7
playsound==1.2.2
gunicorn==20.1.0
python-dotenv==0.19.0 